server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
neo4j==5.28.1
pydantic-settings==2.8.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
